+++ lib/handbook/addon/docs/components/contributor-list/demo-show-less/component.ts
@@ -1,3 +1,0 @@
-import DemoMixin from '../mixin/component';
-
-export default class ContributorListShowLess extends DemoMixin {}