+++ lib/handbook/addon/docs/components/contributor-list/demo-show-more-link/component.ts
@@ -1,3 +1,0 @@
-import DemoMixin from '../mixin/component';
-
-export default class ContributorListMoreLink extends DemoMixin {}