+++ lib/osf-components/addon/components/contributor-list/component.ts
@@ -1,6 +1,5 @@
 import { tagName } from '@ember-decorators/component';
 import { action, computed } from '@ember-decorators/object';
-import { alias } from '@ember-decorators/object/computed';
 import { service } from '@ember-decorators/service';
 import Component from '@ember/component';
 import { task } from 'ember-concurrency';
@@ -9,8 +8,8 @@ import I18N from 'ember-i18n/services/i18n';
 
 import Contributor from 'ember-osf-web/models/contributor';
 import Node from 'ember-osf-web/models/node';
-import Preprint from 'ember-osf-web/models/preprint';
-import Registration from 'ember-osf-web/models/registration';
+import { QueryHasManyResult } from 'ember-osf-web/models/osf-model';
+import Ready from 'ember-osf-web/services/ready';
 import defaultTo from 'ember-osf-web/utils/default-to';
 
 import styles from './styles';
@@ -23,72 +22,49 @@ export interface Contrib {
 }
 
 @tagName('span')
-export default class ContributorList extends Component {
-    layout = layout;
-    styles = styles;
-
-    @service i18n!: I18N;
-    @service store!: DS.Store;
-
-    nodeId?: string;
-    node?: Node | Preprint | Registration;
-
-    page = 1;
-    defaultStep = 3;
-
-    loadingMore: boolean = false;
-    initialLoad: boolean = true;
-
-    showStep: number = defaultTo(this.showStep, 0);
-    contribsPerPage: number = defaultTo(this.contribsPerPage, 10);
-    displayedContribs: number = defaultTo(this.displayedContribs, this.defaultStep);
-
-    showLoading: boolean = defaultTo(this.showLoading, true);
-    darkBall: boolean = defaultTo(this.darkBall, true);
-    userIsBot: boolean = defaultTo(this.userIsBot, navigator.userAgent.includes('Prerender'));
-
-    useLinks?: boolean = defaultTo(this.useLinks, false);
-    useContributorLink?: boolean = defaultTo(this.useContributorLink, false);
-    useShowMoreLink?: boolean = defaultTo(this.useShowMoreLink, false);
-    useShowLess?: boolean = this.userIsBot ? false : defaultTo(this.useShowLess, false);
-
-    showNonBibliographic?: boolean = defaultTo(this.showNonBibliographic, false);
-
-    contributors!: DS.PromiseManyArray<Contributor> &
-                      { meta: { total: number } };
-
-    moreContribs = task(function *(this: ContributorList, more?: boolean) {
-        // we have enough contributors to show, don't request any more.
-        const hiddenContribs = this.contributorList.length - this.displayedContribs;
-        if (!this.userIsBot && (this.rest <= hiddenContribs)) {
-            if (this.useShowStep) {
-                this.incrementProperty('displayedContribs', this.showStep);
-            } else {
-                this.set('displayedContribs', this.contributorList.length);
-            }
-            return;
-        }
-
+export default class ContributorList extends Component.extend({
+    didReceiveAttrs(this: ContributorList) {
         if (!this.node) {
             return;
         }
+        this.loadContributors.perform();
+    },
 
-        // we are running in Prerender, load and show all contributors at once.
+    loadContributors: task(function *(this: ContributorList, more?: boolean) {
+        // Running in Prerender, load and show all contributors at once.
         if (this.userIsBot) {
-            if (!this.hasMore) {
-                this.set('displayedContribs', this.contributorList.length);
-            } else {
-                const allContributors = yield this.node.loadAll('contributors');
-                this.set('contributors', allContributors);
-                this.set('displayedContribs', allContributors.length);
-                this.set('initialLoad', false);
-            }
+            const botBlocker = this.ready.getBlocker();
+
+            const allContributors = yield this.node.loadAll('contributors');
+            this.set('contributors', allContributors);
+            this.set('displayedContribs', allContributors.length);
+
+            botBlocker.done();
             return;
         }
 
-        // Fetch more contributors.
-        this.set('loadingMore', true);
+        const contributors = yield this.node.hasMany('contributors');
+        if (!more && contributors) {
+            this.set('contributors', contributors.value());
+            return;
+        }
 
+        if (more) {
+            // we have enough contributors to show, don't request any more.
+            const hiddenContribs = this.contributorList.length - this.displayedContribs;
+            if (this.rest <= hiddenContribs) {
+                if (this.useShowStep) {
+                    this.incrementProperty('displayedContribs', this.showStep);
+                } else {
+                    this.set('displayedContribs', this.contributorList.length);
+                }
+                return;
+            }
+        }
+
+        const blocker = this.ready.getBlocker();
+
+        // Fetch more contributors.
         const contribs = yield this.node.queryHasMany('contributors', {
             page: more ? this.incrementProperty('page') : this.set('page', 1),
             'page[size]': this.contribsPerPage,
@@ -96,7 +72,6 @@ export default class ContributorList extends Component {
 
         if (more && this.contributors) {
             this.contributors.pushObjects(contribs);
-
             if (this.useShowStep) {
                 this.incrementProperty('displayedContribs', this.showStep);
             } else {
@@ -106,8 +81,34 @@ export default class ContributorList extends Component {
             this.set('contributors', contribs);
         }
 
-        this.set('loadingMore', false);
-    }).restartable();
+        blocker.done();
+    }).restartable(),
+}) {
+    layout = layout;
+    styles = styles;
+
+    @service i18n!: I18N;
+    @service store!: DS.Store;
+    @service ready!: Ready;
+
+    page = 1;
+    defaultStep = 3;
+    userIsBot: boolean = navigator.userAgent.includes('Prerender');
+    contributors!: QueryHasManyResult<Contributor>;
+
+    // Required arguments
+    node!: Node;
+
+    // Optional arguments
+    showStep: number = defaultTo(this.showStep, 0);
+    contribsPerPage: number = defaultTo(this.contribsPerPage, 10);
+    displayedContribs: number = defaultTo(this.displayedContribs, this.defaultStep);
+    showLoading: boolean = defaultTo(this.showLoading, true);
+    dark: boolean = defaultTo(this.dark, true);
+    useContributorLink: boolean = defaultTo(this.useContributorLink, false);
+    useShowMoreLink: boolean = defaultTo(this.useShowMoreLink, false);
+    useShowLess: boolean = defaultTo(this.useShowLess, false);
+    showNonBibliographic: boolean = defaultTo(this.showNonBibliographic, false);
 
     @computed('contributors.[]', 'showNonBibliographic')
     get contributorList(this: ContributorList): Contrib[] {
@@ -126,12 +127,8 @@ export default class ContributorList extends Component {
                 title: c.users.get('familyName') || c.users.get('givenName') || c.users.get('fullName'),
                 id: c.get('unregisteredContributor') ? null : c.users.get('id'),
             }));
-
         return contribs;
     }
-
-    @alias('contributorList.firstObject') first!: Contrib;
-    @alias('contributors.meta.total') numContributors!: number;
 
     @computed('showStep')
     get useShowStep(this: ContributorList): boolean {
@@ -140,6 +137,9 @@ export default class ContributorList extends Component {
 
     @computed('contributors.meta.total', 'displayedContribs')
     get rest(this: ContributorList): number {
+        if (!this.contributors) {
+            return 0;
+        }
         return this.contributors.meta.total - this.displayedContribs;
     }
 
@@ -148,26 +148,19 @@ export default class ContributorList extends Component {
         return this.contributors ? this.contributors.toArray().length < this.contributors.meta.total : undefined;
     }
 
-    @computed('contributors.meta.total', 'showStep', 'displayedContribs', 'useShowLess')
+    @computed('contributors.meta.total', 'showStep', 'displayedContribs', 'useShowLess', 'userIsBot')
     get showLess(this: ContributorList): boolean {
-        return (this.useShowLess as boolean) &&
+        if (!this.contributors) {
+            return false;
+        }
+        return !this.userIsBot && (this.useShowLess as boolean) &&
               (this.displayedContribs > this.defaultStep) &&
             (this.displayedContribs >= this.contributors.meta.total);
     }
 
-    @computed('useLinks', 'useContributorLink')
-    get showContributorLink(this: ContributorList): boolean | undefined {
-        return this.useLinks || this.useContributorLink;
-    }
-
-    @computed('useLinks', 'useShowMoreLink', 'node', 'hasMore')
-    get showMoreLink(this: ContributorList): boolean | undefined {
-        return (this.node || !this.hasMore) && (this.useLinks || this.useShowMoreLink);
-    }
-
     @action
     more(this: ContributorList) {
-        this.get('moreContribs').perform(true);
+        this.get('loadContributors').perform(true);
         return false;
     }
 
@@ -175,13 +168,5 @@ export default class ContributorList extends Component {
     less(this: ContributorList) {
         this.set('displayedContribs', this.defaultStep);
         return false;
-    }
-
-    didReceiveAttrs(this: ContributorList) {
-        if (this.userIsBot && this.initialLoad) {
-            if (this.rest > 0) {
-                this.get('moreContribs').perform(true);
-            }
-        }
     }
 }